requirements.txt
asgiref==3.2.3
Click==7.0
Django==3.0.3
django-click==2.1.0
django-extensions==2.2.8
django-redis-counter==0.1.1
djangorestframework==3.11.0
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.0
psycopg2
uwsgi